+++ lib/src/interface.dart
@@ -1,7 +1,9 @@
 abstract class SystemIdleChecker {
   /// {@template system_idle_initialize}
-  /// Initializes SystemIdle plugin with given time
-  /// If time is not provided the default time is used;
+  /// Initializes the plugin to fire [onIdleStateChanged] when the user is idle for [duration].
+  ///
+  /// This method is idempotent, so you can call it again with a different duration. The default
+  /// duration is 5 minutes.
   /// {@endtemplate}
   Future<void> initialize({required Duration duration});
 
@@ -10,8 +12,17 @@ abstract class SystemIdleChecker {
   /// {@endtemplate}
   Stream<bool> onIdleStateChanged();
 
-  // Returns how long the user has been idle for already.
+  /// {@template system_idle_get_idle_duration}
+  /// Returns how long the user has been idle for already.
+  /// {@endtemplate}
   Future<Duration> getIdleDuration();
+
+  /// {@template system_idle_dispose}
+  /// Releases any resources associated with this object.
+  ///
+  /// Once you call this, this object may not be used again.
+  /// {@endtemplate}
+  void dispose();
 }
 
 class SystemIdleException implements Exception {