+++ lib/src/system_idle_linux.dart
@@ -3,16 +3,25 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
-import 'package:watcher/watcher.dart';
 
 import "interface.dart";
 
-const _idleFilePath = '/tmp/.is_idle';
+// const _idleFilePath = '/tmp/.is_idle';
 const _idleTimePath = '/tmp/idle_time';
 
 class SystemIdleCheckerLinux extends SystemIdleChecker {
   final _controller = StreamController<bool>.broadcast();
+
+  Process? _process;
+  Timer? _timer;
   var lastIsIdle = false;
+
+  @override
+  void dispose() {
+    _process?.kill();
+    _timer?.cancel();
+    _controller.close();
+  }
 
   @override
   Future<void> initialize({required Duration duration}) async {
@@ -23,28 +32,23 @@ class SystemIdleCheckerLinux extends SystemIdleChecker {
     }
 
     try {
+      _process?.kill();
       final execFile = File(Platform.resolvedExecutable);
       final scriptFile = File(p.join(execFile.parent.path, 'idle'));
 
       final idleScript = _generateIdleScript(duration.inSeconds);
       await scriptFile.writeAsString(idleScript);
-      await Process.start('chmod', ['+x', scriptFile.path]);
+      await Process.run('chmod', ['+x', scriptFile.path]);
 
-      Process.run(scriptFile.path, []);
-      FileWatcher(_idleFilePath).events.listen((event) {
-        if (event.type == ChangeType.REMOVE) {
-          _controller.sink.add(false);
-          return;
+      _process = await Process.start(scriptFile.path, []);
+      _timer?.cancel();
+      _timer = Timer.periodic(const Duration(seconds: 1), (timer) async {
+        final idleFor = await getIdleDuration();
+        final isIdle = idleFor > duration;
+        if (lastIsIdle != isIdle) {
+          _controller.sink.add(isIdle);
+          lastIsIdle = isIdle;
         }
-        Timer.periodic(const Duration(seconds: 1), (timer) async {
-          final contents = File(_idleFilePath).readAsStringSync();
-          final isIdle = contents.contains('true');
-
-          if (lastIsIdle != isIdle) {
-            _controller.sink.add(isIdle);
-            lastIsIdle = isIdle;
-          }
-        });
       });
     } catch (e) {
       throw const SystemIdleException(message: "An unknown error occurred");
@@ -90,19 +94,6 @@ idleloop() {
 
         echo \$t > /tmp/.idle_time
         mv /tmp/.input /tmp/.last_input -f
-
-        if [ \$t -ge \$idletime ] && [[ \$a == "2" ]]
-        then
-            echo "user has gone idle"
-            echo "true" > /tmp/.is_idle
-            a=1
-        fi
-        if [ \$t -lt \$idletime ] && [[ \$a == "1" ]]
-        then
-            echo "user has come back from idle"
-            echo "false" > /tmp/.is_idle
-            a=2
-        fi
     done
 }